230428/performence_test/requirements.txt
asgiref==3.6.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
ConfigArgParse==1.5.3
Django==3.2.13
djangorestframework==3.14.0
Flask==2.2.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.15.1
MarkupSafe==2.1.2
msgpack==1.0.5
numpy==1.24.2
pandas==1.5.3
psutil==5.9.4
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==306
pyzmq==25.0.2
requests==2.28.2
roundrobin==0.0.4
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zope.event==4.6
zope.interface==6.0
